+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `Fixed` for any bug fixes
 -   `Security` in case of vulnerabilities
 
+## [UNRELEASED]
+
+### Changed
+
+-   Commands only show when an editor is active
+
 ## [0.5.0] - 2025-12-27
 
 ### Added