+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   Commands only show when an editor is active
+-   Whitespace before and after expressions gets trimmed
 
 ## [0.5.0] - 2025-12-27